requirements.txt
pytorch
torchvision
visdom
